+++ acceleration/opencl/OpenCLDewarp.cpp
@@ -140,7 +140,7 @@ QImage dewarp(
 
 	// Write the source image to device memory.
 	std::array<size_t, 3> const origin{0, 0, 0};
-	std::array<size_t, 3> region{src.width(), src.height(), 1};
+	std::array<size_t, 3> region{(size_t)src.width(), (size_t)src.height(), 1};
 	command_queue.enqueueWriteImage(
 		src_image, CL_FALSE, origin, region, adapted.image.bytesPerLine(), 0,
 		(void*)adapted.image.bits(), &events, &evt
